fix:changes client side to ssr in voluntary page
